requirements.txt
Django==1.7.7
django-bower
django-grappelli
django-filebrowser
git+git://github.com/TND/django-files-widget.git
djangorestframework